replace 'or' operator by 'and'
This apparently results in Elasticsearch requiring *all* words in the
query to occur in a field before it is considered a match, thus "fefe
fnord" won't match the event "Fnord Jahresrückblick" with "fefe" in the
person-field anymore. On the other hand, it prevents "CCC
Jahresrückblick" from occuring when somebody enters "Fnord
Jahresrückblick" as a query.

All in all, this probably makes the search more useful for now.